Getting the front end attached to the backend, unfortunately my attempt to keep the frontend and backend as separate git projects failed.  In theory I could possible create a gradle or maven script that will pull the data from both repos and combine into the complete appliaction, but that is a nice to have, and a post build step.
For now I have combined the both.  The good news is I an ignore the templating provided by Flask, I was fearful this would make the frontend work I done required a significant refactorting. Thankfully it was not the case.  The only item I had to do was move the javasScript to the static folder, and the reference for the url (com.dcu....) in Index.html to point to the static location.

Change-Id: If547ed5b4de5bdb56f4bb89b319082d91a01cb79