Refactoring, separating the functionality that loads the map reduced data form the application.
Added socket-io to the application to prepare for websocket implementation

As part of this commit, a dependency has been placed on the bundle flask-socketio.  Users will be require to run 'pip install flask-socketio' to the version of python.  From with sypder, this is done by (as could not install from conda command)
* in the bottom right panel, where iPython and console tabs exist, select console tab
* In the drop down menu, top right of console tab, open it and select 'Open a command prompt'
* enter pip install flask-socketio

Change-Id: I30b8f00e0f8c374a8ca10b1f15ae0d0364f3dcda