Load data into database through config file and related refactoring
This commit updates the processing on how the data is loaded into the database.  Previously
data was loaded into the database through referencing the location of the data file inline in the
code.  This meant if any user wished to load data into the database, they needed to update
the code, adding the location entry, and rerun the code.

With this commit, the referencing of the location of the data has been moved to a properties
file, dbdata.cfg.  Now, if a user wish to load data into the database, the need only update
the property entry in this file to point to the location of the data

In addtion, a function has been added to ease how the loading of the data is performed.  Now
from the root of the application in pyton all that is needed is to import, the file DatabaseDataLoader
and called the function loadDatabaseData()

Change-Id: I1ed28f0009bf2c34eed869d943a7bf702bddfe36